--- a/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/InternalShared/InterfacesInternal.cs
+++ b/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/InternalShared/InterfacesInternal.cs
@@ -4,19 +4,6 @@ namespace Proto.Promises
 {
     internal static partial class Internal
     {
-        // Abstract class is used instead of interface, because virtual calls on interfaces are twice as slow as virtual calls on classes.
-        internal abstract class ValueContainer
-        {
-            internal abstract void Retain();
-            internal abstract void Release();
-            internal abstract Promise.State GetState();
-            internal abstract Type ValueType { get; }
-            internal abstract object Value { get; }
-
-            internal abstract void ReleaseAndMaybeAddToUnhandledStack(bool shouldAdd);
-            internal abstract void AddToUnhandledStack();
-        }
-
         internal partial interface ITraceable { }
 
         internal interface ILinked<T> where T : class, ILinked<T>