--- a/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/Internal/MergeInternal.cs
+++ b/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/Internal/MergeInternal.cs
@@ -150,7 +150,7 @@ namespace Proto.Promises
                     {
                         if (Interlocked.CompareExchange(ref _valueOrPrevious, valueContainer, null) == null)
                         {
-                            valueContainer.Retain();
+                            _valueOrPrevious = valueContainer.Clone();
                             Handle(ref _waitCount, ref handler, out nextHandler, ref executionScheduler);
                         }
                         if (InterlockedAddWithOverflowCheck(ref _waitCount, -1, 0) == 0)
@@ -166,7 +166,7 @@ namespace Proto.Promises
                         {
                             if (Interlocked.CompareExchange(ref _valueOrPrevious, valueContainer, null) == null)
                             {
-                                valueContainer.Retain();
+                                _valueOrPrevious = valueContainer.Clone();
                                 Handle(ref _waitCount, ref handler, out nextHandler, ref executionScheduler);
                             }
                         }
@@ -195,7 +195,7 @@ namespace Proto.Promises
                         _onPromiseResolved = null;
                         if (_valueContainer != null)
                         {
-                            _valueContainer.Release();
+                            _valueContainer.DisposeAndMaybeAddToUnhandledStack(false);
                             _valueContainer = null;
                         }
 #if PROMISE_DEBUG
@@ -233,7 +233,7 @@ namespace Proto.Promises
                         {
                             if (Interlocked.CompareExchange(ref _valueOrPrevious, valueContainer, null) == null)
                             {
-                                valueContainer.Retain();
+                                _valueOrPrevious = valueContainer.Clone();
                                 Handle(ref _waitCount, ref handler, out nextHandler, ref executionScheduler);
                             }
                             if (InterlockedAddWithOverflowCheck(ref _waitCount, -1, 0) == 0)