--- a/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/ResultContainers.cs
+++ b/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/ResultContainers.cs
@@ -73,7 +73,7 @@ namespace Proto.Promises
         public bool TryGetValueAs<T>(out T value)
         {
             Validate();
-            return Internal.TryGetValue(_valueContainer, out value);
+            return _valueContainer.TryGetValue(out value);
         }