… ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/PromiseTStatic.cs
… ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/Unity/PromiseYieldInstruction.cs