Optimized ValueContainers when `T` is a reference type by only creating one type `ResolveContainer<object>` instead of a type for every `T`.
Optimized ValueContainers for the common case.
Fixed a memory leak with All/Merge/Race/First promises.
Fixed state check in `PromiseYieldInstruction`.

Current master:

```
|         Type |       Method | Pending | Recursion |     Mean |   Error |  StdDev | Allocated |
|------------- |------------- |-------- |---------- |---------:|--------:|--------:|----------:|
|   AsyncAwait | ProtoPromise |    True |       100 | 294.4 us | 5.85 us | 6.96 us |         - |
| ContinueWith | ProtoPromise |    True |       100 | 299.2 us | 1.60 us | 1.49 us |         - |
```

This PR:

```
|         Type |       Method | Pending | Recursion |     Mean |   Error |  StdDev | Allocated |
|------------- |------------- |-------- |---------- |---------:|--------:|--------:|----------:|
|   AsyncAwait | ProtoPromise |    True |       100 | 282.9 us | 1.37 us | 1.14 us |         - |
| ContinueWith | ProtoPromise |    True |       100 | 291.1 us | 1.90 us | 1.78 us |         - |
```